Continue Learning button reusable now, and re-used
