+++ src/components/dashboard/main-content/course-enrollments/course-cards/InProgressCourseCard.jsx
@@ -7,6 +7,8 @@ import { sendTrackEvent } from '@edx/frontend-platform/analytics';
 
 import BaseCourseCard from './BaseCourseCard';
 import { MarkCompleteModal } from './mark-complete-modal';
+import ContinueLearningButton from './ContinueLearningButton';
+
 import Notification from './Notification';
 
 import {
@@ -26,18 +28,11 @@ const InProgressCourseCard = ({
   const [isMarkCompleteModalOpen, setIsMarkCompleteModalOpen] = useState(false);
 
   const renderButtons = () => (
-    <a
-      className="btn btn-outline-primary btn-xs-block"
-      href={linkToCourse}
-      onClick={() => {
-        sendTrackEvent('edx.learner_portal.course.continued', {
-          course_run_id: courseRunId,
-        });
-      }}
-    >
-      Continue Learning
-      <span className="sr-only">for {title}</span>
-    </a>
+    <ContinueLearningButton
+      linkToCourse={linkToCourse}
+      title={title}
+      courseRunId
+    />
   );
 
   const filteredNotifications = notifications.filter((notification) => {
